Refactor to separate file
