+++ event-handler/src/lib/segment-service.ts
@@ -1,14 +1,8 @@
-import { Analytics, TrackParams } from '@segment/analytics-node';
+import { Analytics } from '@segment/analytics-node';
 import { getLogger } from '../utils/logger.utils';
-import {
-  Customer,
-  LineItem,
-  Order,
-  ShippingInfo,
-  TypedMoney,
-} from '@commercetools/platform-sdk';
+import { Customer, Order } from '@commercetools/platform-sdk';
 import { readConfiguration } from '../utils/config.utils';
-import Decimal from 'decimal.js';
+import { buildOrderCompletedTrackEvent } from './segment-event-builder';
 
 const createAnalytics = () => {
   const configuration = readConfiguration();
@@ -82,9 +76,7 @@ export function trackOrderCompleted(order: Order) {
   const analytics = createAnalytics();
 
   try {
-    // https://segment.com/docs/connections/spec/ecommerce/v2/#order-completed
-
-    const event: TrackParams = buildOrderCompletedTrackEvent(order);
+    const event = buildOrderCompletedTrackEvent(order);
 
     analytics.track(event);
 
@@ -96,153 +88,3 @@ export function trackOrderCompleted(order: Order) {
     throw error;
   }
 }
-
-const buildOrderCompletedTrackEvent = (order: Order) => {
-  if (!order.taxedPrice) {
-    throw new Error(`Order ${order.id} is missing taxedPrice`);
-  }
-
-  if (!order.taxedShippingPrice) {
-    throw new Error(`Order ${order.id} is missing taxedShippingPrice`);
-  }
-
-  const products = order.lineItems.map((lineItem, i) => {
-    const imageUrl =
-      lineItem.variant.images && lineItem.variant.images.length > 0
-        ? lineItem.variant.images[0]?.url
-        : undefined;
-
-    return {
-      product_id: lineItem.productId,
-      sku: lineItem.variant.sku,
-      price: getTypedMoneyInCurrencyUnits(getLineItemPrice(lineItem)),
-      quantity: lineItem.quantity,
-      image_url: imageUrl,
-      position: i + 1,
-    };
-  });
-
-  const subTotalCentAmount =
-    order.taxedPrice.totalNet.centAmount -
-    order.taxedShippingPrice.totalNet.centAmount;
-
-  const shippingCentAmount = order.taxedShippingPrice.totalGross.centAmount;
-
-  const subTotalCurrencyUnits = getCentAmountInCurrencyUnits(
-    subTotalCentAmount,
-    order.totalPrice.fractionDigits
-  );
-
-  let discountTotalCents = order.lineItems.reduce((acc, lineItem) => {
-    return (
-      acc +
-      lineItem.discountedPricePerQuantity.reduce(
-        (discountAcc, discountedPricePerQuantity) => {
-          return (
-            discountAcc +
-            discountedPricePerQuantity.discountedPrice.includedDiscounts.reduce(
-              (discountedPriceAcc, discount) => {
-                return (
-                  discountedPricePerQuantity.quantity *
-                    discount.discountedAmount.centAmount +
-                  discountedPriceAcc
-                );
-              },
-              0
-            )
-          );
-        },
-        0
-      )
-    );
-  }, 0);
-
-  discountTotalCents +=
-    order.discountOnTotalPrice?.discountedAmount?.centAmount ?? 0;
-
-  discountTotalCents += getShippingDiscountInCents(order);
-
-  const event: TrackParams = {
-    userId: order.customerId as string, // need either userId or anonymousId
-    anonymousId: order.anonymousId,
-    timestamp: order.createdAt,
-    messageId: `${order.id}-order-completed`,
-    event: 'Order Completed',
-    properties: {
-      email: order.customerEmail,
-      order_id: order.id,
-      total: getTypedMoneyInCurrencyUnits(order.taxedPrice.totalGross), // Subtotal ($) with shipping and taxes added in
-      subtotal: subTotalCurrencyUnits, // subtotal: Order total after discounts but before taxes and shipping
-      // gross discount amount
-      discount: getCentAmountInCurrencyUnits(
-        discountTotalCents,
-        order.totalPrice.fractionDigits
-      ),
-      // gross shipping amount
-      shipping: getCentAmountInCurrencyUnits(
-        shippingCentAmount,
-        order.totalPrice.fractionDigits
-      ),
-      tax:
-        order.taxedPrice.totalTax !== undefined
-          ? getTypedMoneyInCurrencyUnits(order.taxedPrice.totalTax)
-          : undefined,
-      // coupon is a defined as a string in Spec: V2 Ecommerce Events, so just using the first discount code
-      coupon:
-        order.discountCodes && order.discountCodes.length > 0
-          ? order.discountCodes[0].discountCode
-          : undefined,
-      products,
-      currency: order.totalPrice.currencyCode,
-    },
-  };
-  return event;
-};
-
-const getShippingDiscountInCents = (order: Order) => {
-  if (order.shippingMode === 'Multiple') {
-    const shippingDiscountCentAmount = order.shipping.reduce(
-      (acc, shipping) => {
-        return acc + getShippingInfoDiscountInCents(shipping.shippingInfo);
-      },
-      0
-    );
-
-    return shippingDiscountCentAmount;
-  }
-
-  if (!order.shippingInfo) {
-    return 0;
-  }
-
-  return getShippingInfoDiscountInCents(order.shippingInfo);
-};
-
-const getLineItemPrice = (lineItem: LineItem) => {
-  return lineItem.price.discounted
-    ? lineItem.price.discounted.value
-    : lineItem.price.value;
-};
-
-const getShippingInfoDiscountInCents = (shippingInfo: ShippingInfo) => {
-  const discountedCentAmount = shippingInfo.discountedPrice?.value?.centAmount;
-
-  if (!discountedCentAmount) {
-    return 0;
-  }
-
-  return shippingInfo.price.centAmount - discountedCentAmount;
-};
-
-const getTypedMoneyInCurrencyUnits = (money: TypedMoney) => {
-  return getCentAmountInCurrencyUnits(money.centAmount, money.fractionDigits);
-};
-
-const getCentAmountInCurrencyUnits = (
-  centAmount: number,
-  fractionDigits: number
-) => {
-  return new Decimal(centAmount)
-    .div(new Decimal(10).pow(fractionDigits))
-    .toNumber();
-};
